# dozer-api/api.yaml
openapi: 3.0.1
info:
  title: Orchestration
  description: Dozer Api
  version: "1.0"
  contact:
    name: Dozer-Team
    email: api@getdozer.io
    url: "https://getdozer.io"
...
servers:
  - url: "http://localhost:3000"
paths:
...
  /connections/test:
    post:
      summary: ""
      operationId: post-sources-test-connection
      responses:
        "200":
          $ref: "#/components/responses/ConnectionTest"
      description: ""
      requestBody:
        $ref: "#/components/requestBodies/ConnectionTest"
    parameters: []
  /connections:
    post:
      summary: ""
      operationId: post-connections
      requestBody:
        $ref: "#/components/requestBodies/Connection"
      responses:
        "200":
          $ref: "#/components/responses/Connection"
components:
  schemas:
    ConnectionDetails:
      title: ConnectionDetails
      type: object
      properties:
        database:
          type: string
        user:
          type: string
        host:
          type: string
        name:
          type: string
        port:
          type: string
      required:
        - database
        - user
        - host
        - name
    ConnectionAuthentication:
      allOf:
        - $ref: "#/components/schemas/ConnectionDetails"
        - type: object
          properties:
            password:
              type: string
          required:
            - password
      title: ConnectionAuthentication
...
    ConnectionType:
      type: string
      title: ConnectionType
      x-stoplight:
        id: yad8zg2rhqu7c
      enum:
        - postgres
        - snowflake
        - databricks
    ConnectionResponse:
      title: Connection
      allOf:
        - $ref: "#/components/schemas/ConnectionBase"
        - type: object
          properties:
            id:
              type: string
          required:
            - id
    ConnectionRequest:
      title: ConnectionRequest
      allOf:
        - $ref: "#/components/schemas/ConnectionBase"
      x-examples:
        example-1:
          details:
            database: string
            user: string
            host: string
            name: string
          type: postgres
    ConnectionBase:
      title: ConnectionBase
      x-stoplight:
        id: zggzx1lnelwet
      type: object
      properties:
        authentication:
          $ref: "#/components/schemas/ConnectionAuthentication"
        type:
          $ref: "#/components/schemas/ConnectionType"
      required:
        - authentication
        - type
...
    TestConnectionResponse:
      title: TestConnectionResponse
      x-stoplight:
        id: cxwk5w11om32k
      type: object
      properties:
        table_info:
          type: array
          items:
            $ref: "#/components/schemas/TableInfo"
    TableInfo:
      title: TableInfo
      x-stoplight:
        id: fxqjbrkm1kris
      type: object
      properties:
        table_name:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
      required:
        - table_name
    ColumnInfo:
      title: ColumnInfo
      x-stoplight:
        id: p24sf4deu8rrf
      type: object
      properties:
        column_name:
          type: string
        is_nullable:
          type: boolean
        udt_name:
          type: string
        is_primary_key:
          type: boolean
      required:
        - column_name
        - is_nullable
        - udt_name
        - is_primary_key
  securitySchemes: {}
  responses:
    ConnectionTest:
      description: Example response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TestConnectionResponse"
          examples:
            example-1:
              value:
                id: string
        application/xml:
          schema:
            type: object
            properties: {}
...
    Connection:
      description: Example response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ConnectionResponse"
  requestBodies:
...
    ConnectionTest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ConnectionRequest"
        application/xml:
          schema:
            type: object
            properties: {}
    Connection:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ConnectionRequest"
...